Bug fix: A pre HAL 1.2 device cannot write a dynamic temporary
Prior to HAL 1.2, the dimensions of a model output must be specified
either in the Model or in the Request, but in the case of partitioned
compilation and execution, a dynamic temporary (i.e., a step model
output that is not a source model output) will not have its dimensions
specified either in the (step) Model or in the corresponding Request.
The old (buggy) behavior is that we will nevertheless attempt to
execute the Request, which ought to result in an execution-time
validation failure.

Before we implemented support for dynamic temporaries, the behavior
was that if partitioning produces a dynamic temporary, partitioning
fails (which may result in whole-model CPU fallback).  The fix for the
current bug is to restore this behavior EXCEPT that partitioning fails
only if a dynamic temporary is the output of a pre HAL 1.2 device.

TestPartitioning.cpp changes: Add regression tests for this bug fix.
This requires changing the model used by DynamicTemporariesTest to use
only operations available at HAL 1.1.

TestPartitioningRandom.cpp changes:
- Tolerate partitioned compilation failure if there's an operand of
  unknown dimensions and at least one pre HAL 1.2 device.  This is
  similar to how this test behaved before we added support for dynamic
  temporaries (except the device HAL version was irrelevant).
- Check certain properties of a successful partitioned compilation.
- TestDriverV1_[210] should never produce an IPreparedModel of a later
  version than the corresponding IDevice.  Reimplement them in terms
  of HIDL-generated V1_[210]::ADevice, which delegates to an underlying
  IDevice but conforms to the appropriate dynamic type (e.g.,
  V1_2::ADevice can delegate to an underlying V1_3::IDevice, but
  inherits from V1_2::IDevice and ensures that any returned interfaces
  likewise have dynamic types appropriate to the declarations from
  V1_2::IDevice).

Other changes:
- Move the HAL version to API version table from VersionedInterfaces.cpp
  to LegacyHalUtils.h.
- More logging.

Bug: 174851714
Test: NNT_s 1

Change-Id: Id50015cd9a348c7ffb7fa2395ad653e464e73588