Add NDK APIs to query preferred memory alignment and padding.
This CL adds NDK APIs to query the preferred alignment and padding for
input/output buffers or memory object regions. The alignment and padding
is reported at the level of a specific compilation input/output.

The current implementation always returns a fixed value. We will add HAL
APIs for drivers to report the preferences in future updates.

Bug: 179691454
Test: NNT_static
Change-Id: Ib39ab3497b22eb2438f2a1a8754d5ea593852aca