Add IBurst to NN AIDL HAL -- runtime
Burst executions are a sequence of executions of the same prepared model
that occur in rapid succession, such as frames of a camera capture or
successive audio samples. A burst object is used to control a set of
burst executions, and to preserve resources between executions, enabling
executions to have lower overhead.

This CL extends VTS's SamplePreparedModel to minimally implement
configureExecutionBurst and SampleBurst to prevent compilation failures
and pass VTS.

Bug: 180492058
Bug: 177267324
Test: mma
Test: VTS+presubmit
Change-Id: Id72201a69130f12184e1afc2dd74e7b282f55a50