Make NNAPI Memory type more structured
This CL makes NNAPI memory more structured by replacing the pervasive
use of NativeHandle with the new, more specific memory types "Ashmem"
and "MmapFd". It changes the Memory type to be a std::variant of Ashmem
and MmapFd, in addition to (a cleaned up) HardwareBuffer. Finally, the
Memory type also continues to hold the alternative type "NativeHandle"
(now as an "Unknown" memory type) to ensure all previous types of
memory could still be held.

This change also updates the utility and testing code to account for the
changes to the Memory type.

Bug: 183118727
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I7e2fea015a18b5e00d56bef924ebca4449d5b6dd