Allow vendor specific compilation and execution hints
Two new APIs are added to NeuralNetworksExtension.h:
 - ANeuralNetworksCompilation_addExtensionAttribute
 - ANeuralNetworksExecution_addExtensionAttribute

The hints can be associated with the corresponding compilation or
execution objects, through the NNAPI extension mechanism.

Internally, the hints are stored as a list of token-value pairs. The
canonical interface and other internal interfaces have been updated to
pass the hints.

It is the driver's responsiblity to verify the attribute data. And
it is allowed for the driver to ignore the user provided hints.

Bug: 203248587
Test: NeuralNetworksTest_static
Change-Id: I46d012d911d141bd9e9a58aa795feccb75e0b4b0